Moving eth copies
